requirements.txt
tensorflow>=2.12.0
opencv-python>=4.7.0
numpy>=1.23.5
matplotlib>=3.7.1
scikit-learn>=1.2.2
Pillow>=9.5.0
Flask>=2.3.2
flask-cors>=4.0.0
